# rest-api/public/OpenAPISpec.json
{
	"components": {
...
		"schemas": {
			"IntId": {
				"type": "integer",
				"format": "int32",
				"description": "An integer that serves as a record id. Used for primary keys and foreign keys.",
				"minimum": 1
			},
...
			"MediaType": {
				"description": "Media Type",
				"properties": {
					"MediaTypeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"MediaTypeId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Price": {
				"type": "number",
				"format": "float",
				"description": "A price value. Has a precision of 2 significant digits after the decimal point."
			},
...
			"ValidateErrorJSON": {
				"description": "A validation error",
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InvoiceLineResource": {
				"properties": {
					"InvoiceLineId": {
						"$ref": "#/components/schemas/IntId"
					},
					"InvoiceId": {
						"$ref": "#/components/schemas/IntId"
					},
					"TrackId": {
						"$ref": "#/components/schemas/IntId"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					},
					"Quantity": {
						"$ref": "#/components/schemas/IntId"
					}
				},
				"required": [
					"InvoiceLineId",
					"InvoiceId",
					"TrackId",
					"UnitPrice",
					"Quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InvoiceResource": {
				"properties": {
					"InvoiceId": {
						"$ref": "#/components/schemas/IntId"
					},
					"CustomerId": {
						"$ref": "#/components/schemas/IntId"
					},
					"InvoiceDate": {
						"type": "string",
						"format": "date-time"
					},
					"BillingAddress": {
						"type": "string"
					},
					"BillingCity": {
						"type": "string"
					},
					"BillingState": {
						"type": "string"
					},
					"BillingCountry": {
						"type": "string"
					},
					"BillingPostalCode": {
						"type": "string"
					},
					"Total": {
						"$ref": "#/components/schemas/Price"
					},
					"Lines": {
						"items": {
							"$ref": "#/components/schemas/InvoiceLineResource"
						},
						"type": "array"
					}
				},
				"required": [
					"InvoiceId",
					"CustomerId",
					"InvoiceDate",
					"Total",
					"Lines"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InvoiceLineInsertInput": {
				"description": "Invoice Line Insert Input",
				"properties": {
					"TrackId": {
						"$ref": "#/components/schemas/IntId"
					},
					"UnitPrice": {
						"$ref": "#/components/schemas/Price"
					},
					"Quantity": {
						"$ref": "#/components/schemas/IntId"
					}
				},
				"required": [
					"TrackId",
					"UnitPrice",
					"Quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InvoiceInsertInput": {
				"description": "Invoice Insert Input",
				"properties": {
					"CustomerId": {
						"$ref": "#/components/schemas/IntId"
					},
					"InvoiceDate": {
						"type": "string",
						"format": "date-time"
					},
					"BillingAddress": {
						"type": "string"
					},
					"BillingCity": {
						"type": "string"
					},
					"BillingState": {
						"type": "string"
					},
					"BillingCountry": {
						"type": "string"
					},
					"BillingPostalCode": {
						"type": "string"
					},
					"Total": {
						"$ref": "#/components/schemas/Price"
					},
					"Lines": {
						"items": {
							"$ref": "#/components/schemas/InvoiceLineInsertInput"
						},
						"type": "array"
					}
				},
				"required": [
					"CustomerId",
					"InvoiceDate",
					"Total",
					"Lines"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MediaTypeInsertInput": {
				"description": "MediaType Insert Input",
				"properties": {
					"Name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"MediaTypeUpdateInput": {
				"description": "MediaType Update Input",
				"properties": {
					"Name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "Sample REST API built on the Chinook database, for Hasura demo",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "benoit ranque"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/invoices": {
			"get": {
				"operationId": "GetInvoices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/InvoiceResource"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get a set of resources. Can be paginated",
				"tags": [
					"Invoice"
				],
				"security": [],
				"parameters": [
					{
						"description": "how many records should be returned",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "how many records to offset by",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "InsertInvoice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvoiceResource"
								}
							}
						}
					}
				},
				"description": "Create a new instance of the resource. Returns the created resource.",
				"tags": [
					"Invoice"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InvoiceInsertInput",
								"description": "Request Body"
							}
						}
					}
				}
			}
		},
		"/invoices/{InvoiceId}": {
			"get": {
				"operationId": "GetInvoice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvoiceResource"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Get a single resource using the resource id.",
				"tags": [
					"Invoice"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "InvoiceId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		},
		"/mediaTypes": {
			"get": {
				"operationId": "GetMediaTypes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MediaType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get a set of resources. Can be paginated",
				"tags": [
					"MediaType"
				],
				"security": [],
				"parameters": [
					{
						"description": "how many records should be returned",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "how many records to offset by",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "InsertMediaType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MediaType"
								}
							}
						}
					}
				},
				"description": "Create a new instance of the resource. Returns the created resource.",
				"tags": [
					"MediaType"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MediaTypeInsertInput",
								"description": "Request Body"
							}
						}
					}
				}
			}
		},
		"/mediaTypes/{MediaTypeId}": {
			"get": {
				"operationId": "GetMediaType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MediaType"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Get a single resource using the resource id.",
				"tags": [
					"MediaType"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "MediaTypeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateMediaType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MediaType"
								}
							}
						}
					}
				},
				"description": "Update the resource using the resource id. Returns the updated resource.",
				"tags": [
					"MediaType"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "MediaTypeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MediaTypeUpdateInput",
								"description": "Request Body"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMediaType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MediaType"
								}
							}
						}
					}
				},
				"description": "Delete the resource using the resource id. Returns the deleted resource.",
				"tags": [
					"MediaType"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "MediaTypeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}